Adding trait to create symbol table
